Confirmation popup before uninstalling, initial loading screen

+++ src/views/tabs/installed.cpp
@@ -1,5 +1,6 @@
 #include "installed.h"
 #include "installation/manager.h"
+#include "../popup.h"
 
 bool toggleInstalledMod(brls::View* mod_view) {
     InstalledMod* clicked_mod = installed_mods->getInstalledMod(std::stoi(mod_view->getID()));
@@ -19,9 +20,15 @@ bool toggleInstalledMod(brls::View* mod_view) {
 
 Installed::Installed() {
     this->inflateFromXMLRes("xml/tabs/installed.xml");
+
+    // set up uninstallation prompt/inner action
+    confirm_uninstall_popup = new Popup("Are you sure you want to uninstall?");
+    this->addView(confirm_uninstall_popup);
+
     // read from installed items json and populate based on that
     json mem_json_installed = installed_mods->GetMemJson();
     if (mem_json_installed.contains("Installed") && mem_json_installed["Installed"].size() > 0) {
+        CURL_builder curl = CURL_builder();
         for (json entry : mem_json_installed["Installed"]) {
             if (!entry.is_null()) {
 
@@ -29,21 +36,16 @@ Installed::Installed() {
                 if (entry.contains(gb::Fields::Custom::Paths))
                     paths = entry[gb::Fields::Custom::Paths].get<std::vector<std::filesystem::path>>();
 
-                std::string version;
-                if (entry.contains(gb::Fields::AdditionalInfo::AdditionalInfo))
-                    version = entry[gb::Fields::AdditionalInfo::AdditionalInfo][gb::Fields::AdditionalInfo::Version].get<std::string>();
-
                 InstalledMod* m = new InstalledMod({
                     entry[gb::Fields::Name].get<std::string>(), 
                     entry[gb::Fields::Submitter::Submitter][gb::Fields::Name].get<std::string>(), 
-                    version.empty() ? "0.0.0" : version,
                     entry.contains(gb::Fields::Custom::Enabled) ? entry[gb::Fields::Custom::Enabled].get<bool>() : true,
                     entry[gb::Fields::idRow].get<std::string>(), 
                     entry[gb::Fields::Custom::ThumbnailURL].get<std::string>(), 
                     paths
                 });
 
-                this->addInstalledItem(m);
+                this->addInstalledItem(m, &curl);
 
             }
         }
@@ -51,7 +53,7 @@ Installed::Installed() {
 
 }
 
-void Installed::addInstalledItem(InstalledMod* mod) {
+void Installed::addInstalledItem(InstalledMod* mod, CURL_builder* curl) {
     brls::Box* installed_item = (brls::Box*)brls::View::createFromXMLResource("views/installed_item.xml");
 
     // text related elements
@@ -62,16 +64,13 @@ void Installed::addInstalledItem(InstalledMod* mod) {
     brls::Label* installed_item_author = (brls::Label*)installed_item->getView("installed_item_author");
     installed_item_author->setText("Author: " + mod->author);
 
-    brls::Label* installed_item_ver = (brls::Label*)installed_item->getView("installed_item_ver");
-    installed_item_ver->setText("Ver. " + mod->ver);
-
     // thumbnail
 
     if (!NO_IMAGES) {
         brls::Image* installed_item_thumbnail = (brls::Image*)installed_item->getView("installed_item_thumbnail");
         if (!mod->thumbnail_url.empty()) {
-            MemoryStruct img = curl::DownloadToMem(mod->thumbnail_url);
-            //setBrlsImageAsync(mod->thumbnail_url, installed_item_thumbnail);
+            MemoryStruct img = curl::DownloadToMem(mod->thumbnail_url, curl);
+            //brlsImageAsync(mod->thumbnail_url, installed_item_thumbnail);
             installed_item_thumbnail->setImageFromMem((unsigned char*)img.memory, img.size);
         }
     }
@@ -79,8 +78,10 @@ void Installed::addInstalledItem(InstalledMod* mod) {
     // toggle
 
     brls::Image* installed_item_toggle = (brls::Image*)installed_item->getView("installed_item_toggle");
-    if (mod->enabled)
-        installed_item_toggle->setImageFromRes("icon/toggle_on.png");
+    if (!mod->enabled) {
+        installed_item_toggle->setImageFromRes("icon/toggle_off.png");
+        installed_item->setAlpha(0.5);
+    }
     
     // add to installed_mods / register click action
     std::string id = std::to_string(installed_mods->getInstalledModsSize());
@@ -91,13 +92,19 @@ void Installed::addInstalledItem(InstalledMod* mod) {
 
     installed_item->registerClickAction(toggleInstalledMod);
     installed_item->registerAction(
-        "Uninstall", brls::ControllerButton::BUTTON_X, [this, mod, installed_mods, installed_item, installed_box] (brls::View* v) {
-            if (installed_item && mod) {
-                std::string name = mod->name;
-                Manager::UninstallMod(mod); // 'mod' gets freed
-                installed_box->removeView(installed_item);
-                brls::Application::giveFocus(installed_box);
-                brls::Logger::debug("Uninstalled: {}", name);
+        "Uninstall", brls::ControllerButton::BUTTON_X, [this, mod, installed_item, installed_box] (brls::View* v) {
+            if (installed_item && mod && this->confirm_uninstall_popup != nullptr) {
+                this->confirm_uninstall_popup->registerCustomCallback(
+                    [mod, installed_item, installed_box] () {
+                        std::string name = mod->name;
+                        Manager::UninstallMod(mod); // 'mod' gets freed
+                        installed_box->removeView(installed_item);
+                        brls::Application::giveFocus(installed_box);
+                        brls::Logger::debug("Uninstalled: {}", name);
+                    }
+                );
+                this->confirm_uninstall_popup->setVisibility(brls::Visibility::VISIBLE);
+                brls::Application::giveFocus(confirm_uninstall_popup);
             }
             return false;
         }, false, brls::Sound::SOUND_CLICK
@@ -110,6 +117,7 @@ void Installed::addInstalledItem(InstalledMod* mod) {
 void Installed::willAppear(bool resetState) {
     Box::willAppear(resetState);
     setTopText("A -> Toggle | X -> Uninstall");
+    brls::Application::giveFocus(this->getView("installed_box"));
 }
 
 void Installed::willDisappear(bool resetState) {
